Refactor: split redux to give each state its own reducer
