+++ redux/actions/history.js
@@ -1,0 +1,12 @@
+export function saveInj(siteAndTime) {
+    return {
+        type: 'history-save-injection',
+        item: siteAndTime
+    };
+}
+
+export function resetHistory() {
+    return {
+        type: 'history-reset-defaults'
+    };
+}
