+++ redux/actions/sites.js
@@ -1,0 +1,11 @@
+export function nextInjSite() {
+    return {
+        type: 'sites-next-injection-site'
+    };
+}
+
+export function resetSites() {
+    return {
+        type: 'sites-reset-defaults'
+    };
+}
