--- redux/actions/tasks.js
@@ -1,18 +1,0 @@
-export function saveInj(siteAndTime) {
-    return {
-        type: 'tasks-save-injection',
-        item: siteAndTime
-    };
-}
-
-export function nextInjSite() {
-    return {
-        type: 'tasks-next-injection-site'
-    };
-}
-
-export function resetDefaults() {
-    return {
-        type: 'tasks-reset-defaults'
-    };
-}
